Get rid of the busybox init wrapper
Don't think this needs to be written in C is the first problem, portability
is the second given the hardcoded busybox path.

But it's also just largely pointless.
Nobody should want their init to not be named init.
If they do they can write a shebang if they really want to.